Fix type conversion warnings
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@7273 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ include/AnsiCodes.h
@@ -82,7 +82,7 @@ inline std::string stripAnsiCodes(std::string text)
 {
   std::string str = "";
 
-  int length = text.size();
+  int length = (int)text.size();
   for (int i = 0; i < length; i++) {
     if (text[i] == ESC_CHAR) {
       i++;
